transfer reusable functions for search into searchUtils. Delete useless functions
